Fix https://github.com/rpbouman/huey/issues/125
Problem was that a back request would pop the state, which would lead to a request storm of query model updates.
That was not a problem, but we forgot to check the count event passed by the event buffer, so we would repeatedly push the state of intermediate query model states.