Fix https://github.com/rpbouman/huey/issues/412
Issue was that the export dialog contains a progress dialog,which also has a header.
This is significant because in the dragstart of dialog dragging,we check whether the mouse cursor is inside the dialog header.
(and this in turn has to do with preventing drag to interfere with outer pointer events, such as text selection)
The check looked for the first header inside the dialog,and this picked up the header of the progressdialog, not the actual dialog.

For now we moved the progress dialog.
We might want to constrain this a little more in the futre (ie. don't use querySelector, but explicitly scan for child of dialog)