Fix https://github.com/rpbouman/huey/issues/492
This addes settings for the json reader, and introduces logic that allows a failed attempt to read from the datasource to be retried.
The datasource will examine the error message, and might use the error message to adapt or modify its settings, and then retry.

In the case of the json reader returning a '"maximum_object_size" of \d+ bytes exceeded' message, it will extract the '(> \d+ byets)' part from the message, and use that number to retry.
This will repeat until either the number of attempts runs out, or DuckDB runs out of memory, in which case the normal failure reporting kicks in.